Updated Shiitake Door puzzle
Since it's two gates in game, I was considering consolidating the assets
into hard code instead of creating placed instances inside the map and
editing them there... so I've modified the code to accomodate that if I
choose to do so in the future.  For now, it's still instanced as two
gates. But Gate1 is pretty well hard-coded, while Gate2 is the true
instance.